crawler/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiolimiter==1.2.1
aiosignal==1.3.1
APScheduler==3.11.0
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorlog==6.9.0
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
filelock==3.16.1
frozendict==2.4.6
frozenlist==1.5.0
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
jmespath==1.0.1
joblib==1.4.2
lxml==5.3.0
multidict==6.1.0
multitasking==0.0.11
mysql-connector-python==9.1.0
nltk==3.9.1
numpy==2.1.3
packaging==24.2
pandas==2.2.3
parsel==1.9.1
peewee==3.17.8
platformdirs==4.3.6
propcache==0.2.1
Protego==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
queuelib==1.7.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
soupsieve==2.6
tldextract==5.1.3
tqdm==4.67.1
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
w3lib==2.2.1
webencodings==0.5.1
websocket-client==1.8.0
yarl==1.18.3
yfinance==0.2.50
zope.interface==7.2
